Introduced AbstractMenuNode and concrete MenuNode and MultilangMenuNode
